TFP-4708: Rydder opp kode duplisering.
